+++ notification-service/src/main/java/com/spendsmart/notification/service/NotifServiceImpl.java
@@ -19,7 +19,7 @@ public class NotifServiceImpl implements NotifService {
     }
 
     @Override
-    public void sendBudgetAlert(int recipientId, String message, double amount) {
+    public void sendBudgetAlert(Long recipientId, String message, double amount) {
         Notification notification = new Notification();
         notification.setRecipientId(recipientId);
         notification.setType("BUDGET_EXCEEDED");
@@ -31,8 +31,8 @@ public class NotifServiceImpl implements NotifService {
     }
 
     @Override
-    public void sendBulk(List<Integer> recipients, String title, String message) {
-        for (Integer id : recipients) {
+    public void sendBulk(List<Long> recipients, String title, String message) {
+        for (Long id : recipients) {
             Notification n = new Notification();
             n.setRecipientId(id);
             n.setTitle(title);
@@ -43,38 +43,38 @@ public class NotifServiceImpl implements NotifService {
     }
 
     @Override
-    public void markAsRead(int notificationId) {
+    public void markAsRead(Long notificationId) {
         Notification n = repository.findById(notificationId).orElseThrow();
         n.setRead(true);
         repository.save(n);
     }
 
     @Override
-    public void markAllRead(int recipientId) {
+    public void markAllRead(Long recipientId) {
         List<Notification> list = repository.findByRecipientId(recipientId);
         list.forEach(n -> n.setRead(true));
         repository.saveAll(list);
     }
 
     @Override
-    public void acknowledge(int notificationId) {
+    public void acknowledge(Long notificationId) {
         Notification n = repository.findById(notificationId).orElseThrow();
         n.setAcknowledged(true);
         repository.save(n);
     }
 
     @Override
-    public List<Notification> getByRecipient(int recipientId) {
+    public List<Notification> getByRecipient(Long recipientId) {
         return repository.findByRecipientId(recipientId);
     }
 
     @Override
-    public int getUnreadCount(int recipientId) {
+    public int getUnreadCount(Long recipientId) {
         return repository.countByRecipientIdAndIsRead(recipientId, false);
     }
 
     @Override
-    public void deleteNotification(int notificationId) {
+    public void deleteNotification(Long notificationId) {
         repository.deleteById(notificationId);
     }
